requirements.txt
asgiref==3.3.4
attrs==21.2.0
Babel==2.9.1
boto3==1.17.82
botocore==1.20.82
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.11
django-cors-headers==3.7.0
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==3.0.2
django-oscar-api==3.0.0
django-phonenumber-field==3.0.1
django-storages==1.11.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==2.12.0
Faker==8.3.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==20.9
phonenumbers==8.12.23
Pillow==8.2.0
psycopg2==2.8.6
purl==1.6
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
s3transfer==0.4.2
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
swagger-spec-validator==2.7.3
text-unidecode==1.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
